requirements.txt
braceexpand==0.1.7
elasticdeform==0.4.9
ftfy==6.0.3
invoke==1.6.0
mock==4.0.3
msgpack_python==0.5.6
nltk==3.6.5
numpy==1.19.5
pandas==1.2.4
Pillow==8.4.0
pyarrow==6.0.1
pytest==6.2.5
pyzmq==22.3.0
regex==2021.10.23
requests==2.24.0
scikit_learn==1.0.1
setuptools==50.3.1.post20201107
torch==1.9.0
torchaudio==0.10.0
torchvision==0.10.0
tqdm==4.51.0
transformers==4.12.0
wandb==0.12.7
